* [Profiles/Pictures] Implemented a profile picture (i.e. avatar) service. This allows an image to be associated with any record: workers, groups, contacts, email addresses, Virtual Attendants, etc. Previously, this was only supported through the Gravatar plugin, which had some issues (it didn't work on intranets or behind firewalls, workers couldn't set images for users, users were redirected to a third-party service, when the plugin wasn't in use we had no fallback).  When editing a record that supports profile pictures, an "avatar editor" popup is now available.  This allows images to be uploaded or imported by URL, and then zoomed/panned/cropped. Internally, the profile picture service in Cerb works much like Gravatar -- images are requested through the /avatars/<context>/<id> path.  Default images are available per record type (e.g. distinguishing between VAs and workers).  For people-based records (e.g. addys, contexts, workers), there are several silhouettes that are randomly used.  The first time an email-based profile picture (e.g. workers, addys, contacts) is requested and doesn't exist locally, Cerb will check Gravatar for an image for that address and save it locally if found.

+++ features/cerberusweb.core/api/listeners.classes.php
@@ -883,11 +883,18 @@ class ChCoreEventListener extends DevblocksEventListenerExtension {
 		@$context = $event->params['context'];
 		@$context_ids = $event->params['context_ids'];
 		
+		if(empty($context))
+			return;
+		
+		if(!is_array($context_ids) || empty($context_ids))
+			return;
+		
 		// Core
 		DAO_AttachmentLink::removeAllByContext($context, $context_ids);
 		DAO_Calendar::deleteByContext($context, $context_ids);
 		DAO_Comment::deleteByContext($context, $context_ids);
 		DAO_ContextActivityLog::deleteByContext($context, $context_ids);
+		DAO_ContextAvatar::deleteByContext($context, $context_ids);
 		DAO_ContextRecommendation::deleteByContext($context, $context_ids);
 		DAO_ContextLink::delete($context, $context_ids);
 		DAO_CustomFieldset::deleteByOwner($context, $context_ids);
